# Fixed #33532 -- Optimized CaseInsensitiveMapping instantiation for dicts.