@kezabelle 2022-02-21
Updated the patch, which now _doesn't_ special-case in the `__init__` as per the [ticket](https://code.djangoproject.com/ticket/33532#ticket), but gates it behind the `_unpack_items`, which I'd _expected_ to be some of the overhead being saved, but it doesn't look like it substantially changes the change (and seems more likely to be acceptable as a patch):
```
In [3]: %timeit CaseInsensitiveMapping({'Name': 'Jane'})
847 ns ± 14.7 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
In [4]: %timeit HttpHeaders({"HTTP_EXAMPLE": 1})
2.06 µs ± 31.6 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
In [5]: %timeit HttpHeaders({"CONTENT_LENGTH": 1})
2.01 µs ± 27.6 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```