@ngnpope 2022-02-21
For background we recently made changes in bc4c7e5d68e01207268c868932eb14ae4206f02c to optimize stuff here and also consolidated all of the unpacking stuff in `CaseInsensitiveMapping._unpack_items()`. It would be nice if we could avoid moving logic for the `dict` special case back out into `CaseInsensitiveMapping.__init__()`. Also note that bb58d76117513b2af274cb030f67d454fff35218 isn't enough as `ResponseHeaders.__init__()` would also need to be changed, if we were to do this, as it doesn't call `super().__init__()`. Moving this logic out of `_unpack_items()` would be less DRY.

The intent of moving the `dict.items()` call into `_unpack_items()` was also to minimize iterations over the data. Previously in bc4c7e5d68e01207268c868932eb14ae4206f02c there were dictionary comprehensions in the `__init__()` methods for non-mapping iterables after which `dict.items()` had to be called again.

> Turns out I either forgot, or just never knew that `return` in a yielding generator is largely the same as `StopIteration` and that they don't combine in the way I was banking on.

👍🏻 

> Yielding as in [539f82c](https://github.com/django/django/commit/539f82c040b1c068be6cf5b2c8966c7d531918f2) removes much of the gains of the special-casing (costs around `300ns`), so I've reverted back to the original form as the only one to consider or reject (I think...)

So, yes, setting up the generator adds some overhead, but, at least in the case of `HttpHeaders`, your test is atypical. In most cases there are going to be more than one header. So I tried this again with something more realistic for `request.META`.

```python
In [1]: from django.http.request import HttpHeaders
   ...: from django.utils.datastructures import CaseInsensitiveMapping

In [2]: request_meta = {
   ...:     "SERVER_NAME": "localhost",
   ...:     "GATEWAY_INTERFACE": "CGI/1.1",
   ...:     "SERVER_PORT": "8888",
   ...:     "REMOTE_HOST": "",
   ...:     "CONTENT_LENGTH": "",
   ...:     "SCRIPT_NAME": "",
   ...:     "SERVER_PROTOCOL": "HTTP/1.1",
   ...:     "SERVER_SOFTWARE": "WSGIServer/0.2",
   ...:     "REQUEST_METHOD": "GET",
   ...:     "PATH_INFO": "/",
   ...:     "QUERY_STRING": "",
   ...:     "REMOTE_ADDR": "127.0.0.1",
   ...:     "CONTENT_TYPE": "text/plain",
   ...:     "HTTP_HOST": "127.0.0.1:8888",
   ...:     "HTTP_CONNECTION": "keep-alive",
   ...:     "HTTP_CACHE_CONTROL": "max-age=0",
   ...:     "HTTP_SEC_CH_UA": '" Not A;Brand";v="99", "Chromium";v="98"',
   ...:     "HTTP_SEC_CH_UA_MOBILE": "?0",
   ...:     "HTTP_SEC_CH_UA_PLATFORM": '"Linux"',
   ...:     "HTTP_UPGRADE_INSECURE_REQUESTS": "1",
   ...:     "HTTP_USER_AGENT": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/98.0.4758.80 Safari/537.36",
   ...:     "HTTP_ACCEPT": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
   ...:     "HTTP_SEC_FETCH_SITE": "none",
   ...:     "HTTP_SEC_FETCH_MODE": "navigate",
   ...:     "HTTP_SEC_FETCH_USER": "?1",
   ...:     "HTTP_SEC_FETCH_DEST": "document",
   ...:     "HTTP_ACCEPT_ENCODING": "gzip, deflate, br",
   ...:     "HTTP_ACCEPT_LANGUAGE": "en-GB,en-US;q=0.9,en;q=0.8",
   ...:     "HTTP_COOKIE": "csrftoken=ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+=",
   ...:     "wsgi.input": object(),
   ...:     "wsgi.errors": object(),
   ...:     "wsgi.version": (1, 0),
   ...:     "wsgi.run_once": False,
   ...:     "wsgi.url_scheme": "http",
   ...:     "wsgi.multithread": True,
   ...:     "wsgi.multiprocess": False,
   ...:     "wsgi.file_wrapper": object(),
   ...: }
```

659d2421c7adbbcd205604002d521d82d6b0b465 (main):

```python
In [3]: %timeit HttpHeaders(request_meta)
13 µs ± 16.6 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

bb58d76117513b2af274cb030f67d454fff35218:

```python
In [3]: %timeit HttpHeaders(request_meta)
12.3 µs ± 20.5 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

It still has an impact. You mentioned the overhead of comparing `dict` against `Mapping` going through `_abc_instancecheck`. We could just attempt to avoid that with the following:

```diff
diff --git a/django/utils/datastructures.py b/django/utils/datastructures.py
index b5858b8076..63cbaf743c 100644
--- a/django/utils/datastructures.py
+++ b/django/utils/datastructures.py
@@ -327,7 +327,8 @@ class CaseInsensitiveMapping(Mapping):
 
     @staticmethod
     def _unpack_items(data):
-        if isinstance(data, Mapping):
+        # Explicitly test for dict first as the common case for performance.
+        if isinstance(data, (dict, Mapping)):
             yield from data.items()
             return
         for i, elem in enumerate(data):
```

Which did give me the following:

```python
In [6]: %timeit HttpHeaders(request_meta)
12.6 µs ± 30 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

But I can't promise this will be consistent. Sometimes it seemed to make no difference.

Being a micro-optimisation, it'd be interesting to know whether this actually makes a dent in the normal processing of a request as a whole before we go ahead with the change.

----

As an aside, I noticed that `ResponseHeaders._convert_to_charset()` could possibly be improved by reducing the number of calls to `isinstance()`.  Perhaps something like:

```diff
diff --git a/django/http/response.py b/django/http/response.py
index e40f2d169b..a94519a872 100644
--- a/django/http/response.py
+++ b/django/http/response.py
@@ -42,16 +42,20 @@ class ResponseHeaders(CaseInsensitiveMapping):
         `charset` must be 'ascii' or 'latin-1'. If `mime_encode` is True and
         `value` can't be represented in the given charset, apply MIME-encoding.
         """
-        if not isinstance(value, (bytes, str)):
-            value = str(value)
-        if (isinstance(value, bytes) and (b"\n" in value or b"\r" in value)) or (
-            isinstance(value, str) and ("\n" in value or "\r" in value)
+        if isinstance(value, bytes):
+            type_ = bytes
+        else:
+            type_ = str
+            if not isinstance(value, str):
+                value = str(value)
+        if (type_ is bytes and (b"\n" in value or b"\r" in value)) or (
+            type_ is str and ("\n" in value or "\r" in value)
         ):
             raise BadHeaderError(
                 "Header values can't contain newlines (got %r)" % value
             )
         try:
-            if isinstance(value, str):
+            if type_ is str:
                 # Ensure string is valid in given charset
                 value.encode(charset)
             else:
```

I wonder whether we could also delay the check that results in `BadHeaderError` until after the `.decode()` to remove one branch. 🤔